Continued porting code to new data structures, everything but DCResistivity compiles and associated tests run, still need to finish data transforms, first tests indicate no significant problems

+++ trunk/jif3D/MT/ReadWriteTitanData.cpp
@@ -215,7 +215,7 @@ namespace jif3D
             ExIndices.resize(nmeas * nfreq);
             EyIndices.resize(nmeas * nfreq);
             HIndices.resize(nmeas * nfreq);
-            for (int ifr = 0; ifr < nfreq; ++ifr)
+            for (size_t ifr = 0; ifr < nfreq; ++ifr)
               {
                 ind_shift = nmeas * ifr;
                 for (size_t i = 0; i < nmeas; ++i)
